single v3 contract

+++ contracts/src/V3DataSync.sol
@@ -16,7 +16,7 @@ interface IUniswapV3Pool {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint8 feeProtocol,
+            uint32 feeProtocol,
             bool unlocked
         );
     function ticks(int24 tick)
